Separate bands calculation from plotting
A `BandStructure` utility class is introduced to encapsulate all band structure data. Internal and AiiDA mechanics are adjusted w.r.t this class. In the case of AiiDA, the band calculation step returns an instance of this class, which contains all the necessary information to reproduce the band structure plot.